cash saves to db with taxes & account

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -18,7 +18,7 @@ class CreateUsersTable extends Migration
             $table->string('name');
             $table->string('email')->unique();
             $table->string('password');
-            $table->integer('amount')->nullable();
+            $table->float('amount')->nullable();
             $table->rememberToken();
             $table->timestamps();
         });
